% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checklist_to_df}
\alias{checklist_to_df}
\title{checklist_to_df}
\usage{
checklist_to_df(check_list)
}
\arguments{
\item{check_list}{: Results of structure/schema check.}
}
\value{
Dataframe of all check_list information
}
\description{
Tidy results of checks into df to display in app
}
