% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{empty_data_dict}
\alias{empty_data_dict}
\title{empty_data_dict}
\usage{
empty_data_dict(upload_dataset)
}
\arguments{
\item{upload_dataset}{: Dataframe of uploaded dataset csv file.}
}
\value{
Dataframe of partially filled data dictionary, if no data dictionary uploaded.
}
\description{
Create empty data dict template (if no data dict uploaded)
}
