% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{first_char_letter}
\alias{first_char_letter}
\title{first_char_letter}
\usage{
first_char_letter(string)
}
\arguments{
\item{string}{: String to test}
}
\value{
String of "Pass" or "Fail"
}
\description{
Test if variable name starts with a letter.
}
