% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{library.dynam.unload}
\alias{library.dynam.unload}
\title{library.dynam.unload}
\usage{
library.dynam.unload(
  chname,
  libpath,
  verbose = getOption("verbose"),
  file.ext = .Platform$dynlib.ext
)
}
\value{
Unloads all libraries not included in R startup
}
\description{
used only for package version of app
}
