% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{render_check_table}
\alias{render_check_table}
\title{render_check_table}
\usage{
render_check_table(structure_checks, schema_checks)
}
\arguments{
\item{structure_checks}{: Result of structure checks.}

\item{schema_checks}{: Result of schema checks.}
}
\value{
Single dataframe of combined structure and schema check results.
}
\description{
Show results of structure_checks in shiny app
}
