% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{startApp}
\alias{startApp}
\title{startApp}
\usage{
startApp(loadReqPackages = TRUE)
}
\arguments{
\item{loadReqPackages}{: Should the function load required packages on user's behalf? Default is `TRUE`.}
}
\value{
Opens shiny app inside default web browser.
}
\description{
Single functiont to run shiny app (will load required packages by default)
}
