% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.odcCheck}
\alias{summary.odcCheck}
\title{summary.odcCheck}
\usage{
\method{summary}{odcCheck}(validate_results)
}
\arguments{
\item{validate_results}{: Results from structure and schema checks.}
}
\value{
Prints out report for the checks.
}
\description{
Prints out report of checks in console. Not used in shiny app.
}
