% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{system.file}
\alias{system.file}
\title{system.file}
\usage{
system.file(..., package = "base", lib.loc = NULL, mustWork = FALSE)
}
\value{
Path to package (if found)
}
\description{
Find whether package is installed. Only used in package version of app
}
