% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{validate_odc}
\alias{validate_odc}
\title{validate_odc}
\usage{
validate_odc(
  dataset = NULL,
  datadic = NULL,
  dataset_path = NULL,
  datadic_path = NULL,
  str_checks,
  sch_checks
)
}
\arguments{
\item{dataset}{: Dataset dataframe in global environment}

\item{datadic}{: Data dictionary dataframe in global environment}

\item{dataset_path}{: Path to dataset csv}

\item{datadic_path}{: Path to data dictionary csv}

\item{str_checks}{: Default is "all"}

\item{sch_checks}{: Default is "all"}
}
\value{
List of structure and schema check results.
}
\description{
Single (wrapper) function to run both structure & schema checks. Only structure checks are conducted if datadic are not present.
}
