% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{validate_schema}
\alias{validate_schema}
\title{sci = all, tbi = else (only difference is tbi doesn't use "min_vars")
min_vars = minimal_var = vars that must be present in dataset/data dic
wrangle other_symbols_df
Value = TRUE when other symbols detected in var name}
\usage{
validate_schema(dataset, datadic, sch_checks = "all")
}
\description{
sci = all, tbi = else (only difference is tbi doesn't use "min_vars")
min_vars = minimal_var = vars that must be present in dataset/data dic
wrangle other_symbols_df
Value = TRUE when other symbols detected in var name
}
