% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{var_other_symbols}
\alias{var_other_symbols}
\title{var_other_symbols}
\usage{
var_other_symbols(string)
}
\arguments{
\item{string}{: String to test}
}
\value{
String of True if non-accepted symbols present in variable name, Fail if string ONLY contains accepted characters.
}
\description{
Test if var name contains non-accepted symbols (NOT A-z0-9_.)
}
